backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
chromadb==0.4.24
numpy<2.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
rank-bm25==0.2.2
fugashi==1.3.0
unidic-lite==1.0.8
